glamor-egl: Added new function to create image for a pixmap.
This function is used to support dri2. In the underlying
driver, it will create a buffer object for a given pixmap.
And then call this api to create a egl image from that
buffer object, and then bind that image to a texture, and
then bind that texture to the pixmap.

Normally, this pixmap's content is shared between a dri2
client and the x server.

Signed-off-by: Zhigang Gong <zhigang.gong@linux.intel.com>