Track all picture's drawable pict format.
Even if a picture's pixmap is a pure in memory pixmap, we still need
to track its format. The reason is we may need to upload this drawable
to texture thus we need to know its real picture format.

As to the MACRO to check whether a pixmap is a picture, we should
check whether the priv is non-NULL before we touch its field.

Signed-off-by: Zhigang Gong <zhigang.gong@linux.intel.com>