glamor_download_sub_pixmap_to_cpu: New function to download subregion.
Prepare to optimize the fallback path. We choose the important
rendering pathes to optimzie it by using shader. For other pathes,
we have to fallback. We may continue to optimize more pathes in
the future, but now we have to face those fallbacks.

The original fallback is very slow and will download/upload the whole
pixmap. From this commit, I will refine it to just download/upload
needed part.

Signed-off-by: Zhigang Gong <zhigang.gong@linux.intel.com>